Moved interfaces to controllers & services package
